refactor: move types to file for cartAPI
