+++ src/services/Cart/cartAPI.types.ts
@@ -1,0 +1,18 @@
+export interface Cart {
+  id: string;
+  version: number;
+  lineItems: LineItem[];
+}
+
+export interface LineItem {
+  id: string;
+  productId: string;
+  name: string;
+  quantity: number;
+  price: {
+    value: {
+      centAmount: number;
+      currencyCode: string;
+    };
+  };
+}
